Unify CNN and MLP evaluators

+++ src/datasets/dataset.py
@@ -18,7 +18,7 @@ class Dataset(data.Dataset):
 
     def __init__(self, X, y):
         self.X = torch.tensor(X, dtype=torch.float32)
-        self.y = torch.tensor(y, dtype=torch.int64)
+        self.y = torch.tensor(y, dtype=torch.float32)
 
     def __len__(self):
         return len(self.X)
